working remove gradient
