+++ packages/utility/package.json
@@ -40,6 +40,7 @@
         "mime-lite": "^1.0.3",
         "normalize-url": "^8.0.1",
         "strip-ansi": "^7.1.0",
+        "tailwind-merge": "^3.3.1",
         "tldts": "^6.1.82",
         "uuid": "^11.1.0"
     }
